Major reorganization. New Utils.py and Scraper.py files for utilities and the main program, respectfully.
